+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 1.1 (unreleased)
 ----------------
 
+- Remove auto-increment column from ocid_cell table and make the
+  radio, mcc, mnc, lac, cid combination the primary key. Also optimize the
+  column types of the lac and cid fields.
+
 - Update to latest versions of amqp, configparser, cornice, kombu, protobuf,
   pytz and Sphinx.